move market data tabs component to folder
