+++ apps/ui/stratify-ui/app/(screens)/app/markets/MarketDataTabs/MarketDataTabs.test.tsx
